feature(mysql): Attempted to add MySQL
Unfortunately linode (very recently) has stopped users from creating new
instances of managed database services. To this end I wanted to add the
cluster to k8s - however this was just as much of a pain as the last
time I recall (crds with finalizers that never execute, insufficient
resources in k8s), so I'd rather wait a while and give it a swing when more
time is avaiable.

For now I will setup the mysql server manually (it's not hard) and
populate credentials required through pulumi. Instructions: https://www.linode.com/docs/guides/install-and-configure-mysql-on-ubuntu-22-04/